combine yaml test functionality into single module
